% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Means}
\alias{Means}
\title{Means of Things}
\usage{
Means(data, x, y)
}
\arguments{
\item{data}{Data set}

\item{x}{Character column}

\item{y}{Numeric column}
}
\description{
Returns means of numeric columns by character groups
}
\examples{
Means(deliveries1,Date,Tip)
}
