% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SubPerPlot}
\alias{SubPerPlot}
\title{Subtotal vs Tip Percentage Scatterplot (Payment Type points)}
\usage{
SubPerPlot(data, x, y, z, title)
}
\arguments{
\item{data}{Data set}

\item{x}{Subtotal column of data set}

\item{y}{Tip Percentage column of data set}

\item{z}{Payment Type column of data set}

\item{title}{Title of the graph}
}
\description{
This function returns a ggplot point graph
}
\examples{
SubPerPlot(Sunday,Sun_Subtotal,Sun_Percentage,Sun_Payment,"Sunday Subtotals and Tip Percentage")
}
