% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SubTipPlot}
\alias{SubTipPlot}
\title{Subtotal vs Tip Scatterplot (Payment Type points)}
\usage{
SubTipPlot(data, x, y, z, title)
}
\arguments{
\item{data}{Data set}

\item{x}{Subtotal column of data set}

\item{y}{Tip column of data set}

\item{z}{Payment Type column of data set}

\item{title}{Title of the graph}
}
\description{
This function returns a ggplot point graph
}
\examples{
SubTipPlot(Tuesday,Tues_Subtotal,Tues_Tip,Tues_Payment,"Tuesday Subtotals and Tips")
}
