% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Sums}
\alias{Sums}
\title{Sums of Things}
\usage{
Sums(data, x, y)
}
\arguments{
\item{data}{Data set}

\item{x}{Character column}

\item{y}{Numeric column}
}
\description{
Sums of numeric columns by character groups
}
\examples{
Sums(Friday,Date,Tip)
}
