% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TotalPerPlot}
\alias{TotalPerPlot}
\title{Order Total vs Tip Percentage Scatterplot (Payment Type points)}
\usage{
TotalPerPlot(data, x, y, z, title)
}
\arguments{
\item{data}{Data set}

\item{x}{Order Total column of data set}

\item{y}{Tip Percentage column of data set}

\item{z}{Payment Type column of data set}

\item{title}{Title of the graph}
}
\description{
This function returns a ggplot point graph
}
\examples{
TotalPerPlot(Friday,Fri_Order_Total,Fri_Percentage,Fri_Payment,"Friday Order Totals and Tip Percentage")
}
