% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TotalTipPlot}
\alias{TotalTipPlot}
\title{Order Total vs Tip Scatterplot (Payment Type points)}
\usage{
TotalTipPlot(data, x, y, z, title)
}
\arguments{
\item{data}{Data set}

\item{x}{Order Total column of data set}

\item{y}{Tip column of data set}

\item{z}{Payment Type column of data set}

\item{title}{Title of the graph}
}
\description{
This function returns a ggplot point graph
}
\examples{
TotalTipPlot(deliveries1,Order_Total,Tip,Payment,"Order Totals and Tips")
}
