% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{WhatDay}
\alias{WhatDay}
\title{What Day Is It?}
\usage{
WhatDay(x)
}
\arguments{
\item{x}{Date within the data}
}
\description{
Function to return the Day of the week for a Date
}
\examples{
WhatDay("03_09_21")
}
